// src/const/enums/tags.ts
export const enum Tags {
    REACT = "React",
    JS = "Java Script",
    TS = "Type Script",
    NODE = "Node",
    TAILWIND = "Tailwind",
    NEXT = "Next.js",
    MONGO = "Mongo DB",
    EXPRESS = "Express",
    VITE = "Vite",
    CSS = "Css Vanilla",
    ASTRO = "Astro",
    WEB_COMPONENTS = "Web Component",
}